+++ lib/src/widgets/base_number_field.dart
@@ -45,7 +45,7 @@ class BaseNumberField<T> extends StatefulWidget {
   final Icon? icon;
   final Icon? prefixIcon;
   final Icon? suffixIcon;
-  final void Function()? onDeleted;
+  final void Function()? onCleared;
   final TextInputType? keyboardType;
 
   final TextStyle? style;
@@ -76,7 +76,7 @@ class BaseNumberField<T> extends StatefulWidget {
     required this.icon,
     required this.prefixIcon,
     required this.suffixIcon,
-    required this.onDeleted,
+    required this.onCleared,
     required this.keyboardType,
     required this.style,
     required this.textInputAction,
@@ -222,7 +222,7 @@ class _BaseNumberFieldState<T> extends State<BaseNumberField<T>> {
             icon: widget.icon,
             prefixIcon: widget.prefixIcon,
             suffixIcon: widget.suffixIcon,
-            onDeleted: widget.onDeleted,
+            onCleared: widget.onCleared,
             hasInput: _controller.text.isNotEmpty,
             keyboardType: widget.keyboardType,
           ),