+++ lib/src/widgets/optional_string_field.dart
@@ -21,7 +21,7 @@ class OptionalStringField extends StatefulWidget {
   final Icon? icon;
   final Icon? prefixIcon;
   final Icon? suffixIcon;
-  final void Function()? onDeleted;
+  final void Function()? onCleared;
   final TextInputType? keyboardType;
 
   final TextStyle? style;
@@ -44,7 +44,7 @@ class OptionalStringField extends StatefulWidget {
     this.icon,
     this.prefixIcon,
     this.suffixIcon,
-    this.onDeleted,
+    this.onCleared,
     this.keyboardType,
     this.style,
     this.textInputAction,
@@ -112,7 +112,7 @@ class _OptionalStringFieldState extends State<OptionalStringField> {
             icon: widget.icon,
             prefixIcon: widget.prefixIcon,
             suffixIcon: widget.suffixIcon,
-            onDeleted: widget.onDeleted,
+            onCleared: widget.onCleared,
             hasInput: _controller.text.isNotEmpty,
             keyboardType: widget.keyboardType,
           ),